Add Catch framework for testing embedding support and C++-side features
At this point, there is only a single test for interpreter basics.

Apart from embedding itself, having a C++ test framework will also
benefit the C++-side features by allowing them to be tested directly.

The current situation with the extension module and pytest is that
all assertion need to made in Python. This makes testing C++ interface
features awkward since the majority of the logic is on the C++ side.