Port test_eval from pytest to Catch
Testing the Python C++ interface is easier this way because there is
no need to pass data back to Python just for an assertion. Having
everything in C++ makes it easier to write more detailed test with
more assertions. Test failure reports are also more accurate and point
directly to lines in the C++ source.